rename deinit func to cleanup
this was changed because:

deinit always felt mandatory to me, meaning it would
feel wrong to not call it. There are plenty of times
where the code instantiates a statement just to do
some work, in those moments, the statement must not
be deinitialized, otherwise it would free memory that
it does not own. Besides that, i think cleanup is also
more explicit, like, it should only be called if it owns memory.

deinit always felt self sufficient, meaning it doesn't
depend on anything, so passing an allocator to it
felt wrong to me.